Move computationally expensive dataset def to separate action
